backend/requirements.txt
# Core Framework
fastapi>=0.110.0
uvicorn>=0.25.0
starlette>=0.37.0

# Database
motor>=3.3.0
pymongo>=4.5.0

# Authentication & Security
bcrypt>=4.1.0
passlib>=1.7.4
python-jose>=3.5.0
PyJWT>=2.8.0
cryptography>=41.0.0

# Validation & Serialization
pydantic>=2.5.0
email-validator>=2.1.0
python-multipart>=0.0.6

# HTTP & Networking
httpx>=0.25.0
requests>=2.31.0
aiohttp>=3.9.0

# Environment & Config
python-dotenv>=1.0.0

# Image Processing
pillow>=10.0.0

# RSS & Web Scraping
feedparser>=6.0.0
beautifulsoup4>=4.12.0

# Utilities
python-dateutil>=2.8.0

# Rate Limiting
slowapi>=0.1.9

# Email (Resend)
resend>=0.7.0

# Firebase Push Notifications (using HTTP API - no SDK needed)
# PyJWT already included for auth, used for Firebase JWT generation

--extra-index-url https://d33sy5i8bnduwe.cloudfront.net/simple/
redis>=5.0.0
openai>=1.0.0
gunicorn>=21.0.0
cloudinary>=1.36.0
anthropic>=0.40.0
